libinput: remove unneeded patch
The log_msg symbol conflict was fixed in libevdev commit e4c315fc1603 (Rename
symbols leaking from static library to avoid name clashes, 2014-08-18) that
was included in version 1.3 of libevdev. Remove the patch that used to fix the
issue of log_msg symbol conflict with libevedev, which is a direct dependency
of libinput.

Cc: Romain Naour <romain.naour@openwide.fr>
Signed-off-by: Baruch Siach <baruch@tkos.co.il>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>