batman-adv: new package
This package introduces batman-adv, a kernel module implementation of
the B.A.T.M.A.N. IV and V mesh network routing protocols.

While batman-adv exists in the mainline kernel tree, it can also be
built as an external out-of-tree module. This package adds the
flexibility to chose a more up to date version of the module than exists
in the official tree, and also allows for compilation against kernels
without batman-adv in-tree support.

https://www.open-mesh.org/projects/batman-adv/

Signed-off-by: Christian Stewart <christian@paral.in>
[Thomas:
 - remove "default n", since it's the default
 - fix indentation of Config.in help text
 - license is GPLv2, not just GPL
 - remove variable BATMAN_ADV_MAKE_OPTS, name it directly
   BATMAN_ADV_MODULE_MAKE_OPTS as this is what is expected by the kernel
   module infrastructure.]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>