configs: new configuration for NXP i.MX51 EVK
Add a new configuration for NXP i.MX51 EVK based on U-boot 2016.05
and kernel 4.6.1.

U-boot 2016.05 needs the patch c510f2e436008 ("video: ipu_common: fix build
error") that is already in mainline to fix an IPU build error.

We can remove this patch in the future when we switch to U-boot 2016.07.

Signed-off-by: Fabio Estevam <festevam@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>