xdriver_xf86-video-imx-viv: bump version to 5.0.11-p8.4
Remove upstreamed patch.
Add patch from Yocto to not use git for versioning.

This patch is based on the Yocto equivalent:
https://github.com/Freescale/meta-fsl-arm/commit/06a0574
https://github.com/Freescale/meta-fsl-arm/commit/076df90

The following xorg.conf was used in order to force the use of the
vivante module:
https://github.com/Freescale/meta-fsl-arm/blob/master/recipes-graphics/xorg-xserver/xserver-xf86-config/mx6/xorg.conf

Note that the X server must be started with the "noreset" option to
avoid crashes when closing X apps.
Tested with:
 # X -noreset &
 # cd /usr/share/examples/viv_samples/vdk/
 # DISPLAY=:0 ./tutorial7

Signed-off-by: Andrew Webster <awebster@arcx.com>
[Thomas: add upstream reference in the patch, as suggested by Gary.]

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>