libsemanage: bump to version 2.5
The 0001-execption-lib-path-fix.patch is also no longer needed, as the
new version of libsemanage includes the fix provided by the patch.
As such, the patch was removed, and 0002 was renamed to 0001.

Audit was added as a dependency as the new version of libsemanage
will fail to compile searching for audit.h

Signed-off-by: Adam Duskett <Aduskett@gmail.com>
Reviewed-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
[Thomas: add the select on audit + propagate the dependencies.]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>