gdb: add dependency on host-texinfo
Just like binutils, when gdb is fetched from git, it wants to
re-generate its documentation. We were trying to avoid that by patching
the Makefiles, but it wasn't working in all situations.

So, we simply add a dependency on host-texinfo when gdb is fetched from
git, and remove our hackhish work-around.

Fixes:

  http://autobuild.buildroot.net/results/80e63edc009d23cb939880a6f5ae2f220f07a96a/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Reviewed-by: Romain Naour <romain.naour@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>