package/qt5webkit: needs host-python
The Qt5Webkit build system hard-codes calls to 'python', but only really
supports python2.

We fix that similarly as we did in nodejs, by adding a python->python2
symlink early in the PATH when configuring/building Qt5Webkit, and
adding host-python in the dependency list.

Fixes:
    http://autobuild.buildroot.org/results/ea4/ea4dd00ee585f9f9ea1927b8b5fd9035e6733420/
    http://autobuild.buildroot.org/results/323/323ba220fae70d94c81bab9a348be844e90ace91/
    http://autobuild.buildroot.org/results/6c2/6c20c9e222f94e6b70b97b87e1393dd16da2718c/

Also slightly reorganise the dependency list, now that it no longer fits
on a single line.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>