lldpd: bump version to 0.9.4
With this new version, the generation of atom-glue fails with the
following error:

    Making all in src/lib
    make[3]: Entering directory '/mnt/data/src/sfl/zodiac/buildroot/output/build/lldpd-0.9.4/src/lib'
      GEN      atom-glue.c
    arm-buildroot-linux-uclibcgnueabi-cpp.br_real: fatal error: too many input files
    compilation terminated.
    arm-buildroot-linux-uclibcgnueabi-cpp.br_real: fatal error: too many input files
    compilation terminated.
    make[3]: *** [Makefile:899: atom-glue.c] Error 1

There is an upstream patch pushed after 0.9.4 which fixes the
corresponding Makefile.am, included in this commit.

Note that since the provided tarball ships the related Makefile.in file,
we need to tell Buildroot to autoreconfigure the package.

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>
[Thomas: add comment about AUTORECONF=YES in .mk file.]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>